firefox-aurora: Implement launcher that sets up environment
Checks if desktop environment is Wayland and sets appropriate variables to use Wayland

Sets MOZ_APP_LAUNCHER to make crash reporter restart the launcher instead of firefox-bin
and also resolves the browser creating a buggy desktop file when it's trying to set itself
as the default browser. Fixes the browser not recognizing that it's the default when our desktop
file has been selected in the system settings.

Signed-off-by: Namelesswonder <Namelesswonder@users.noreply.github.com>