firefox-aurora: Create native messaging manifest symlink for DE browser integration if necessary
Fixes KDE Plasma Integration and GNOME Shell Integration extensions not working.

The respective integration extensions on the browser side will still have to be manually installed.
https://addons.mozilla.org/en-US/firefox/addon/gnome-shell-integration/
https://addons.mozilla.org/en-US/firefox/addon/plasma-integration/

This functionality is dependent on gnome-browser-connector or plasma-browser-integration being installed on the system.

Also change test operand from == to = as the former is a bashism.

Signed-off-by: Namelesswonder <Namelesswonder@users.noreply.github.com>